requirements.txt
fastapi
uvicorn
SQLAlchemy
databases
asyncpg
passlib[bcrypt]
python-jose
pydantic
alembic
psycopg2